chore: moving schemas into own files
